refactor: extract the presentation subproject
