+++ application/src/main/java/jp/skypencil/application/task/TaskApplicationService.java
@@ -1,8 +1,11 @@
 package jp.skypencil.application.task;
 
+import java.util.Optional;
+import java.util.UUID;
 import java.util.function.Predicate;
 import java.util.stream.Stream;
 import jp.skypencil.domain.model.Task;
+import jp.skypencil.domain.model.TaskId;
 import jp.skypencil.domain.model.TaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,6 +22,20 @@ public class TaskApplicationService {
   @Autowired
   TaskApplicationService(TaskRepository repository) {
     this.repository = repository;
+  }
+
+  public TaskData create(String subject) {
+    Task task = new Task(TaskId.create(), subject, false);
+    repository.save(task);
+    return new TaskData(task);
+  }
+
+  public Optional<TaskData> find(UUID id) {
+    return repository.find(new TaskId(id)).map(TaskData::new);
+  }
+
+  public Stream<TaskData> listAll() {
+    return repository.list().map(TaskData::new);
   }
 
   public Stream<TaskData> listToDo() {
