+++ presentation/build.gradle
@@ -1,0 +1,14 @@
+apply plugin: 'org.springframework.boot'
+apply plugin: 'io.spring.dependency-management'
+
+apply from: "$rootDir/gradle/publish.gradle"
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation project(':application')
+    compileOnly 'com.google.code.findbugs:jsr305:3.0.2'
+    compileOnly "org.immutables:value:${immutablesVersion}:annotations"
+    annotationProcessor "org.immutables:value:${immutablesVersion}"
+    testImplementation "com.google.guava:guava:${guavaVersion}"
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+}
