+++ settings.gradle
@@ -1,3 +1,4 @@
 include ':domain'
 include ':application'
 include ':infrastructure'
+include ':presentation'
